chore: repo hygiene cleanup (Phase 1)
WHY: Remove xmake build system artifacts, stale binary artifacts, and
     unused conan dependencies that accumulated since the CMake migration.
WHAT:
- Remove 27 xmake binary build artifacts from git index (build/)
- Remove xmake.lua, xmake_enhanced.lua, xmake_minimal.lua from index
- Remove src/fs/extent.hpp (moved to include/xinim/fs/extent.hpp)
- Remove third_party/limine/limine_protocol.h (moved to third_party/limine/)
- Remove src/kernel/net_driver.hpp (moved to include/xinim/net/net_driver.hpp)
- Remove .xmake/ and *.xmake patterns from .gitignore (xmake gone)
- Remove libsodium/1.0.19 from conanfile.py (vendored, not Conan-linked)
- Add archive/legacy/xmake/README.md explaining archived xmake contents

HOW: git rm --cached for indexed files, edit .gitignore and conanfile.py

+++ userland/shell/xinim-sh/builtins.cpp
@@ -103,7 +103,7 @@ int builtin_exit(Command* cmd) {
     }
 
     // Clean up jobs
-    for (int i = 0; i < MAX_JOBS; i++) {
+    for (size_t i = 0; i < MAX_JOBS; i++) {
         if (g_shell.jobs[i].pgid != 0 &&
             g_shell.jobs[i].state == JobState::RUNNING) {
             fprintf(stderr, "xinim-sh: There are running jobs\n");
